OpenWithAI/requirements.txt
pystray
Pillow
keyboard
pypdf